+++ UniversityAdmission/UniversityAdmission.Domain/Dto/DtoForRequest/ApplicantCountForSpecialityDto.cs
@@ -1,0 +1,7 @@
+﻿namespace UniversityAdmission.Domain.Dto.DtoForRequest;
+
+public class ApplicantCountForSpecialityDto
+{
+    public int SpecialtyId { get; set; }
+    public int ApplicantCount { get; set; }
+}
