+++ UniversityAdmission/UniversityAdmission.Domain/Dto/DtoForRequest/ApplicantWithMaxScoreWithSpecialty.cs
@@ -1,0 +1,10 @@
+﻿using UniversityAdmission.Domain.Class;
+
+namespace UniversityAdmission.Domain.Dto.DtoForRequest;
+
+public class ApplicantWithMaxScoreWithSpecialty
+{
+    public Applicant Applicant { get; set; }
+    public double MaxScore { get; set; }
+    public int SpecialtyId { get; set; }
+}
