+++ UniversityAdmission/UniversityAdmission.Domain/Dto/DtoForRequest/TopFiveApplicantDto.cs
@@ -1,0 +1,9 @@
+﻿using UniversityAdmission.Domain.Class;
+
+namespace UniversityAdmission.Domain.Dto.DtoForRequest;
+
+public class TopFiveApplicantDto()
+{
+    public Applicant Applicant { get; set; }
+    public double TotalScore { get; set; }
+}
